Create a base class for LiquidLegionsV2Mill and ReachOnlyLiquidLegionsV2Mill.
