+++ src/main/proto/wfa/measurement/internal/duchy/protocol/BUILD.bazel
@@ -137,6 +137,7 @@ proto_library(
     deps = [
         ":liquid_legions_sketch_parameter_proto",
         ":liquid_legions_v2_noise_config_proto",
+        ":liquid_legions_v2_proto",
         "//src/main/proto/wfa/measurement/internal/duchy:crypto_proto",
         "//src/main/proto/wfa/measurement/internal/duchy:differential_privacy_proto",
         "//src/main/proto/wfa/measurement/internal/duchy/config:protocols_setup_config_proto",
