+++ src/test/kotlin/org/wfanet/measurement/duchy/db/computation/ReachOnlyLiquidLegionsSketchAggregationV2ProtocolEnumStagesDetailsTest.kt
@@ -21,8 +21,8 @@ import org.junit.runners.JUnit4
 import org.wfanet.measurement.duchy.db.computation.ReachOnlyLiquidLegionsSketchAggregationV2Protocol.EnumStages.Details
 import org.wfanet.measurement.internal.duchy.ComputationStageDetails
 import org.wfanet.measurement.internal.duchy.computationStageDetails
+import org.wfanet.measurement.internal.duchy.protocol.LiquidLegionsSketchAggregationV2Kt.ComputationDetailsKt.computationParticipant
 import org.wfanet.measurement.internal.duchy.protocol.ReachOnlyLiquidLegionsSketchAggregationV2
-import org.wfanet.measurement.internal.duchy.protocol.ReachOnlyLiquidLegionsSketchAggregationV2Kt.ComputationDetailsKt.computationParticipant
 import org.wfanet.measurement.internal.duchy.protocol.ReachOnlyLiquidLegionsSketchAggregationV2Kt.computationDetails
 import org.wfanet.measurement.internal.duchy.protocol.ReachOnlyLiquidLegionsSketchAggregationV2Kt.stageDetails
 import org.wfanet.measurement.internal.duchy.protocol.ReachOnlyLiquidLegionsSketchAggregationV2Kt.waitSetupPhaseInputsDetails
